+++ deep-print-processor/src/commonMain/kotlin/com/bradyaiello/deepprint/DeepPrintSymbolProcessorProvider.kt
@@ -6,8 +6,10 @@ import com.google.devtools.ksp.processing.SymbolProcessorProvider
 
 class DeepPrintSymbolProcessorProvider : SymbolProcessorProvider {
     override fun create(environment: SymbolProcessorEnvironment): SymbolProcessor {
+        val indent = (environment.options["indent"] ?: "4").toInt()
         return DeepPrintProcessor(
-            codeGenerator = environment.codeGenerator
+            codeGenerator = environment.codeGenerator,
+            indent = indent,
         )
     }
 }
