+++ test-project/build.gradle.kts
@@ -37,8 +37,6 @@ dependencies {
     implementation(project(":external-module"))
     implementation(project(":deep-print-processor"))
     ksp(project(":deep-print-processor"))
-//    implementation("io.github.brady-aiello:deep-print-processor:0.1.0")
-//    ksp(project("io.github.brady-aiello:deep-print-processor:0.1.0"))
     testImplementation(Testing.Junit.jupiter)
     testImplementation(KotlinX.datetime)
 }
